requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.2
Pillow==9.2.0
psycopg2==2.9.4
sqlparse==0.4.3
